requirements.txt
yt-dlp
google-genai
python-slugify
requests
beautifulsoup4
python-dotenv
pytest
